pyproject.toml
[tool.poetry]
name = "exako"
version = "0.1.0"
description = ""
authors = ["Gabriel Lustosa <lustosaki2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-ninja = "^1.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pillow = "^10.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
blue = "^0.9.1"
ruff = "^0.5.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"
ipython = "^8.26.0"
pytest-django = "^4.8.0"
freezegun = "^1.5.1"

[tool.ruff]
exclude = [".venv", "migrations"]

[tool.ruff.format]
quote-style = "single"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
extend_skip = ["migrations"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
DJANGO_SETTINGS_MODULE = "exako.settings.test"

[tool.blue]
extend-exclude = "(migrations/)"

[tool.taskipy.tasks]
lint = "ruff check . --fix && blue --check . --diff"
format = "ruff format .  && isort ."
run = "python manage.py runserver"
pre_test = "task format"
test = "pytest -s --cov=. -vv"
post_test = "coverage html"

